[clang] fix assertion failure when consuming editor placeholders in some declaration contexts
Currently we are using ConsumeToken() while recovering from seeing an
editor placeholder in external declarations. This causes problems when
the token preceeding the placeholder is a special token (one that needs
to be balanced).

Currently this change calls ConsumeAnyToken() instead which fixes the
assertion failure.

Before this can be merged:
 - [ ] If the token is special, be sure to properly handle it.